backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
aiohttp>=3.8.0
supabase==1.0.4
GeoAlchemy2==0.14.1
Pillow>=11.2.1
email-validator
sentry-sdk[fastapi]==1.40.6
